Bug fixes and refactoring
Fix MQTT JSON output, apply batch limit when writing to Recent layer, plus a start at refactoring (method extractions to begin to cut some bulk out of the monolithic main program). Tested but did not resolve: This program cannot be used to write directly to the Feature Service endpoints of an Analytics IoT created Layer. For that, either write to a normal ArcGIS Online feature service, or use an MQTT broker.